Make openslide an extra
